refactor(e2e): remove offline mode handling and related checks from tests
