+++ .github/workflows/run-appium-e2e-workflow.yml
@@ -258,14 +258,6 @@ jobs:
         run: echo "pattern=${{ inputs.test_suite_tag }}" >> "$GITHUB_OUTPUT"
         shell: bash
 
-      - name: Verify Appium runner connectivity (Android)
-        if: ${{ inputs.platform == 'android' }}
-        run: yarn tsx scripts/e2e/verify-appium-runner-connectivity.mjs android
-        env:
-          ANDROID_AVD_NAME: appium_smoke_avd
-          ANDROID_EMULATOR_CI_CORES: '4'
-          ANDROID_SERIAL: ${{ env.ANDROID_SERIAL }}
-
       - name: Verify Appium runner connectivity (iOS)
         if: ${{ inputs.platform == 'ios' }}
         run: yarn tsx scripts/e2e/verify-appium-runner-connectivity.mjs ios
