+++ tests/flows/wallet.flow.ts
@@ -48,7 +48,6 @@ import OnboardingInterestQuestionnaireView from '../page-objects/Onboarding/Onbo
 import ExperienceEnhancerBottomSheet from '../page-objects/Onboarding/ExperienceEnhancerBottomSheet';
 import { fetchProductionFeatureFlags } from '../performance/feature-flag-helper';
 import { ExistingUserSheetSelectorsIDs } from '../../app/components/Views/Notifications/PushNotificationOnboarding/ExistingUserSheet/ExistingUserSheet.testIds';
-import { assertNotOnOfflineModeScreen } from './offline-mode.flow';
 const logger = createLogger({
   name: 'WalletFlow',
 });
@@ -516,18 +515,13 @@ export const loginToAppPlaywright = async (
 ): Promise<void> => {
   const { scenarioType = 'login' } = options;
 
-  try {
-    await PlaywrightAssertions.expectElementToBeVisible(
-      asPlaywrightElement(LoginView.container),
-      {
-        description: 'Login view container',
-        timeout: 45_000,
-      },
-    );
-  } catch (error) {
-    await assertNotOnOfflineModeScreen('while waiting for login screen');
-    throw error;
-  }
+  await PlaywrightAssertions.expectElementToBeVisible(
+    asPlaywrightElement(LoginView.container),
+    {
+      description: 'Login view container',
+      timeout: 45_000,
+    },
+  );
   await PlaywrightAssertions.expectElementToBeVisible(
     asPlaywrightElement(LoginView.passwordInput),
     {
